% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{build_op_url}
\alias{build_op_url}
\title{Build operations url}
\usage{
build_op_url(api, scheme, host, base_path, op_def, par_values)
}
\arguments{
\item{scheme}{http or https}

\item{host}{host name with port (delimited by ":")}

\item{base_path}{base path, defined in api specification}

\item{op_def}{a single operation definition}

\item{par_values}{parameter values in a list}
}
\description{
Build operations operation url for specified parameter values
}
\seealso{
\code{\link{get_operation_definitions}}
}
\keyword{internal}
