% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{default_arguments}
\alias{default_arguments}
\alias{set_default_args}
\alias{set_default_args_call}
\alias{set_default_args_list}
\title{Set default arguments}
\usage{
set_default_args(f, ...)

set_default_args_list(f, arguments)

set_default_args_call(f_call)
}
\arguments{
\item{f}{function}

\item{...}{Parameters with default values}

\item{arguments}{A named list of arguments names and values}

\item{f_call}{A function call}
}
\value{
A function with new defaults on arguments
}
\description{
Use this functions to simplify operation and schema functions with default
arguments
}
\keyword{internal}

