% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{get_operation_definitions}
\alias{get_operation_definitions}
\title{Get Operations Definitions}
\usage{
get_operation_definitions(api, path = NULL)
}
\arguments{
\item{api}{API object}

\item{path}{(optional) filter by path}
}
\description{
A list of operations definitions organized by operationId field
}
\keyword{internal}

