% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{get_par_values}
\alias{get_par_values}
\title{Get parameter values}
\usage{
get_par_values(fun, env)
}
\arguments{
\item{fun}{A function (to be passed to formals)}

\item{env}{environment}
}
\value{
A named list of parameters
}
\description{
Returns parameters, passed to a function with default parameters
}
\keyword{internal}

