% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{get_parameters}
\alias{get_parameters}
\title{Get Parameters}
\usage{
get_parameters(api, api_parameters)
}
\arguments{
\item{api}{API definition}

\item{parameters_def}{A parameters from API operations definition}
}
\description{
Extract all parameters from parameters definition as a list
In case of reference to schema, use the schema.
}
\keyword{internal}
