% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{get_schema_graphviz_dot}
\alias{get_schema_graphviz_dot}
\title{Get schema graphviz dot (experimental in this version)}
\usage{
get_schema_graphviz_dot(api, schema, schema_name = NULL,
  rankdir = "LR", gv_attrs = "")
}
\arguments{
\item{api}{Api object}

\item{schema}{A schema definition from api (a list object)}

\item{schema_name}{A name of root schema}
}
\description{
Create a graphviz presentation of schema structure (experimental)
}
\keyword{internal}
