% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{get_schemas}
\alias{get_schemas}
\title{Get schemas}
\usage{
get_schemas(api)
}
\arguments{
\item{api}{Api object}
}
\value{
A list of functions
}
\description{
Returns a list of functions with arguments from API schemas. Elements are
named by schema names, each function returns a named list.
}
