% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.rapi_api}
\alias{print.rapi_api}
\title{Print API object}
\usage{
\method{print}{rapi_api}(x, ...)
}
\arguments{
\item{x}{API object}

\item{...}{further arguments passed to or from other methods}
}
\description{
Print API object
}
\keyword{internal}
