% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.rapi_operation}
\alias{print.rapi_operation}
\title{Print Operation}
\usage{
\method{print}{rapi_operation}(x, ...)
}
\arguments{
\item{x}{Operation}

\item{...further}{arguments passed to or from other methods}
}
\description{
Print Operation
}
\keyword{internal}
