% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print_wrap}
\alias{print_wrap}
\title{Print Operation}
\usage{
print_wrap(text, ...)
}
\arguments{
\item{text}{character(1) to be formatted to the width of the user
screen, with additional arguments \code{...} passed to
\code{strwrap()}.}
}
\description{
Print Operation
}
\keyword{internal}
