% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{result_handlers}
\alias{content_or_message}
\alias{content_or_stop}
\alias{content_or_warning}
\alias{result_handlers}
\title{Simple functions to handle httr response}
\usage{
content_or_stop(x)

content_or_warning(x)

content_or_message(x)
}
\arguments{
\item{x}{A response object from httr functions (see \link{httr} documentation)}
}
\value{
Content of http response
}
\description{
When creating operations from api one can define
how the response from http should be handled.
These functions can be used for simple result handling.
}
\examples{
\dontrun{
operations <- get_operations(api, handle_response = content_or_stop)
}
}

