% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{get_message_body}
\alias{get_message_body}
\title{Message body}
\usage{
get_message_body(x)
}
\arguments{
\item{x}{A list}
}
\description{
Transform a list to http request message body
}
\keyword{internal}
