% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{get_operation_definitions}
\alias{get_operation_definitions}
\title{Get Operations Definitions}
\usage{
get_operation_definitions(api, path = NULL)
}
\arguments{
\item{api}{API object}

\item{path}{(optional) filter by path}
}
\description{
Get a list of operations definitions from API specification
}
\details{
Operations are parsed from `paths`` element for every path and every action
inside path. Operation name is set to `operationId` from each action.

See also specification \url{http://swagger.io/specification/#operationObject}
}
\keyword{internal}

