% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations.R
\name{get_operations}
\alias{get_operations}
\title{Create Operation Functions}
\usage{
get_operations(api, .headers = NULL)
}
\arguments{
\item{api}{API object (see \code{\link{get_api}})}

\item{.headers}{Optional headers passed to \code{\link[httr]{add_headers}}}
}
\value{
A list of functions.
}
\description{
Creates a list of functions from API operations definition.
Names in a list are operationIDs from API.
}

